Add hook for asking location permission
